requirements.txt
pytest==6.2.5
pyenchant==3.2.2
termcolor==1.1.0